% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{and}
\alias{and}
\title{and}
\usage{
and(x, y)
}
\arguments{
\item{x}{The first logical input.}

\item{y}{The second logical input.}
}
\value{
A logical input containig the resutls of x & y.
}
\description{
Simple function returns true if both of the inputs are true, otherwise it returns false. Equivallent to
the '&' symbol in R & most programming languages. Throws an error if both inputs aren't logical.
}
\examples{
and(c(FALSE,FALSE,TRUE,TRUE),c(FALSE,TRUE,FALSE,TRUE))
}
