% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{check_bool}
\alias{check_bool}
\title{check_bool}
\usage{
check_bool(x)
}
\arguments{
\item{x}{The case checked variable}
}
\description{
Helper function checls whether the input is a logical or not, stops the run if it's not a logical.
}
\examples{
check_bool(c(TRUE,TRUE))
}
