% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{check_length}
\alias{check_length}
\title{check_length}
\usage{
check_length(x, y)
}
\arguments{
\item{x}{The first input}

\item{y}{Thes second input}
}
\description{
Helper functions asserts that both inputs are of the same length.
}
\examples{
check_length(c(1:3),c(1:3))
}
