% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{implies}
\alias{implies}
\title{implies}
\usage{
implies(x, y)
}
\arguments{
\item{x}{The first logical input.}

\item{y}{The second logical input.}
}
\value{
A logical input containig the resutls of x => y.
}
\description{
Simple function returns false only if x is true and y is false , otherwise it returns true.
Throws an error if both inputs aren't logical.
}
\examples{
implies(c(FALSE,FALSE,TRUE,TRUE),c(FALSE,TRUE,FALSE,TRUE))
}
