% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{inhibit}
\alias{inhibit}
\title{inhibit}
\usage{
inhibit(x, y)
}
\arguments{
\item{x}{First logical input.}

\item{y}{Second logical input.}
}
\value{
A logical vector that conatins X and Not Y.
}
\description{
A function that resutls in x but not y (oreder matters).
True if X is true and y is false, otherwise returns false.
Throws errors with non logical inputs & different length inputs.
}
\examples{
inhibit(c(FALSE,FALSE,TRUE,TRUE),c(FALSE,TRUE,FALSE,TRUE))
}
