% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{nand}
\alias{nand}
\title{nand}
\usage{
nand(x, y)
}
\arguments{
\item{x}{The first logical input.}

\item{y}{The second logical input.}
}
\value{
A logical input containig the resutls of !(x&y).
}
\description{
Simple function returns false if both of the inputs are true, otherwise it returns true.
Throws an error if both inputs aren't logical.
}
\examples{
nand(c(FALSE,FALSE,TRUE,TRUE),c(FALSE,TRUE,FALSE,TRUE))
}
