% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{not}
\alias{not}
\title{not}
\usage{
not(x)
}
\arguments{
\item{x}{The logical to be inverted.}
}
\value{
A logical vector, the inverse of the logical parameter x.
}
\description{
Simple function represents the not logic gate & an equivallent for the
 '!' sign in R & most preogramming languages. Throws an error if the input isn't a logical variable.
}
\examples{
not(c(TRUE,FALSE))
}
