% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{null}
\alias{null}
\title{null}
\usage{
null(x)
}
\arguments{
\item{x}{A number indicates the length of returned vector}
}
\description{
Simple function returns a vector of FALSE's of a given length
}
\examples{
null(3)
}
