% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{on}
\alias{on}
\title{on}
\usage{
on(x)
}
\arguments{
\item{x}{A number indicates the length of returned vector}
}
\description{
Simple function returns a vector of TRUE's of a given length
}
\examples{
identity(3)
}
