% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{or}
\alias{or}
\title{or}
\usage{
or(x, y)
}
\arguments{
\item{x}{The first logical input.}

\item{y}{The second logical input.}
}
\value{
A logical input containig the resutls of x | y.
}
\description{
Simple function returns true if one of the inputs is true, otherwise it returns false. Equivallent to
the '|' symbol in R & most programming languages. Throws an error if both inputs aren't logical.
}
\examples{
or(c(FALSE,FALSE,TRUE,TRUE),c(FALSE,TRUE,FALSE,TRUE))
}
