% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{transefer}
\alias{transefer}
\title{transefer}
\usage{
transefer(x)
}
\arguments{
\item{x}{Input}
}
\value{
The same entered input
}
\description{
The simplest gate. Returns it's input without any modification, usually transefer gates
are used to amplify the signal incoming. So in boolean algerbra terms a transefer gate is neutral.
}
\examples{
transefer(TRUE)
}
