% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{truth_table}
\alias{truth_table}
\title{truth_table}
\usage{
truth_table()
}
\value{
A data frame of two variables x & y, contains all possible logical combinations of two logical variables.
}
\description{
A function that creates a general table of cases to test with two variables on several logic gates.
Isn't expected to throw an error.
}
