% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{xor}
\alias{xor}
\title{xor}
\usage{
xor(x, y)
}
\arguments{
\item{x}{The first logical input.}

\item{y}{The second logical input.}
}
\value{
A logical input containig the resutls of x!=y.
}
\description{
Simple function returns true if the two inputs are not of same value, otherwise it returns false.
Throws an error if both inputs aren't logical.
}
\examples{
xor(c(FALSE,FALSE,TRUE,TRUE),c(FALSE,TRUE,FALSE,TRUE))
}
